requirements.txt
aiogram==2.25.1
psycopg2-binary==2.9.6
python-dotenv==0.21.0